+++ src/main/webapp/components/nav-bar/index.js
@@ -1,0 +1,3 @@
+import NavBar from './nav-bar'
+import NavigationBarContext from './nav-bar-context'
+export { NavBar as default, NavigationBarContext }
