+++ src/main/webapp/components/nav-bar/nav-bar-context.js
@@ -1,0 +1,3 @@
+import React from 'react'
+const NavigationBarContext = React.createContext()
+export default NavigationBarContext
