+++ src/main/webapp/components/query-manager/query-manager.stories.tsx
@@ -6,7 +6,7 @@ import useState from '../../@storybook/use-state'
 import sampleQueries from '../../sample-data/sample-queries.json'
 import QueryEditor from '../query-editor'
 const stories = storiesOf('Query Manager', module)
-import { NavigationBarContext } from '../../nav-bar-context'
+import { NavigationBarContext } from '../nav-bar'
 
 import { button } from '@storybook/addon-knobs'
