--- src/main/webapp/nav-bar-context.js
@@ -1,2 +1,0 @@
-import React from 'react'
-export const NavigationBarContext = React.createContext()
